Implemented new solution for Requests in VR Module
