+++ src/components/vr_cooperation/src/command/on_service_deactivated_notification.cc
